+++ Photo Club Hub/Documentation/ReleaseNotes.md
@@ -2,7 +2,7 @@
 
 CLUBS AND MUSEUMS SCREEN
 * Added 10 clubs for a northern district of the Netherlands. They are for now at Level 1. Thanks to AH.
-* Renamed IndividueelBO to PersoonlijkBO (to match Fotobond convention)
+* Renamed IndividueelBO to Persoonlijk16 (to match Fotobond convention: "Persoonlijke Leden")
 * TODO: fix scrolling of Clubs and Museums screen (on iPhone only?). <<<<
 
 DATA MAINTENANCE